refactor: move dht code from mod.rs to client.rs
